// src/api/models/productResponse.ts
/**
 * Generated by orval v7.17.2 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */

export interface ProductResponse {
  id?: number;
  name?: string;
  categoryId?: number;
  price?: number;
  originalPrice?: number;
  description?: string;
  mainImage?: string;
  images?: string;
  stock?: number;
  sales?: number;
  status?: number;
  createdAt?: string;
  updatedAt?: string;
}