// src/api/models/pageProductResponse.ts
/**
 * Generated by orval v7.17.2 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import type { ProductResponse } from './productResponse';
import type { SortObject } from './sortObject';
import type { PageableObject } from './pageableObject';

export interface PageProductResponse {
  totalPages?: number;
  totalElements?: number;
  size?: number;
  content?: ProductResponse[];
  number?: number;
  sort?: SortObject;
  first?: boolean;
  last?: boolean;
  numberOfElements?: number;
  pageable?: PageableObject;
  empty?: boolean;
}